final commit

+++ client/src/Components/Admin/Dashboard/Dashboard.jsx
@@ -32,11 +32,23 @@
     }, []); 
 
     const deleteuser = (id) => {
-      deleteUser(id).then(() => {
+      // Display a confirmation dialog using SweetAlert
+      Swal.fire({
+        title: 'Are you sure?',
+        text: 'You won\'t be able to revert this!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.isConfirmed) {
          
-         fetchUserData()
-        
-        
+          deleteUser(id).then(() => {
+           
+            fetchUserData();
+          });
+        }
       });
     };
     return (
